refactor: Move DevBundleUtils
Nothing relates to the frontend build there
