+++ flow-server/src/main/java/com/vaadin/flow/server/frontend/TaskCopyNpmAssetsFiles.java
@@ -34,6 +34,7 @@ import java.util.function.Predicate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.vaadin.flow.internal.DevBundleUtils;
 import com.vaadin.flow.internal.FrontendUtils;
 
 import static com.vaadin.flow.server.Constants.VAADIN_WEBAPP_RESOURCES;
