+++ flow-server/src/main/java/com/vaadin/flow/server/frontend/TaskCopyTemplateFiles.java
@@ -26,6 +26,7 @@ import java.util.HashSet;
 import java.util.Set;
 
 import com.vaadin.flow.component.dependency.JsModule;
+import com.vaadin.flow.internal.DevBundleUtils;
 import com.vaadin.flow.internal.FrontendUtils;
 import com.vaadin.flow.internal.Template;
 import com.vaadin.flow.server.Constants;
