+++ flow-server/src/test/java/com/vaadin/flow/internal/DevBundleUtilsTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.server.frontend;
+package com.vaadin.flow.internal;
 
 import java.io.File;
 import java.io.IOException;
@@ -25,8 +25,6 @@ import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
-
-import com.vaadin.flow.internal.StringUtil;
 
 public class DevBundleUtilsTest {
