+++ flow-server/src/test/java/com/vaadin/flow/server/frontend/TaskCopyNpmAssetsFilesTest.java
@@ -36,6 +36,7 @@ import org.mockito.Mockito;
 import org.slf4j.LoggerFactory;
 
 import com.vaadin.flow.di.Lookup;
+import com.vaadin.flow.internal.DevBundleUtils;
 import com.vaadin.flow.server.frontend.scanner.ClassFinder;
 import com.vaadin.flow.server.frontend.scanner.FrontendDependenciesScanner;
